refactor(domains): extract the Index Manager body into a shared panel (GH #1543)
The per-domain Index Manager was a Modal-only component. Its form + save now
live in DomainIndexPanel, so it renders both inline (a new "Index Files" tab on
the tenant Web Domain page) and inside the existing row-menu Modal, which now
delegates to the panel with footer={null}. The panel re-syncs from its source
on change (dropping the old open-guarded effect) so an inline pane that stays
mounted can't clobber a concurrent save.

This is the first of the per-domain action extractions for the Web Domain page;
Index goes first because its modal is shared by the admin row menu too, so it
proves the delegate pattern keeps that path working. The remaining tenant panes
(Document root, Domain options, Rewrite rules) follow in the next commit.
